ai-image-fh/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pillow
huggingface_hub
requests
